Segundo Commit
Avanzado un poco.
Hay salto simple, apto para probar el giro de cámara en diferentes puntos Y.

Giro de cámara no funciona bien.
