+++ Prueba Movimiento 2D/Assets/Animations/Base Torre - Empty.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e93287cf91deeab49ad58f381552f467
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
