+++ Prueba Movimiento 2D/Assets/Animations/Cara1.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: df3f276c5d30d2746905016db1338ced
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
