+++ Prueba Movimiento 2D/Assets/Animations/Cara2.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4b4db29aa9f757e44b38562dd0f4e592
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
