+++ Prueba Movimiento 2D/Assets/Animations/Cara3.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5ae86f733e4996b48ab6ab45fbc53b34
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
