+++ Prueba Movimiento 2D/Assets/Animations/Cara4.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ad4a519e3b7a4424787cc2d6c83f7156
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
