+++ Prueba Movimiento 2D/Assets/Animations/Giro1-2.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 238fd5019185ac44188ae969b36a3dd1
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
