+++ Prueba Movimiento 2D/Assets/Animations/Giro2-3.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e1c9e75592e314448b806901a585203
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
