+++ Prueba Movimiento 2D/Assets/Animations/Giro3-4.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0480960a20e86a846b39bcdb5f0c0790
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
