+++ Prueba Movimiento 2D/Assets/Animations/Giro4-1.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 70bb5efc68033a64e989c052b2b0e060
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
