+++ Prueba Movimiento 2D/Assets/Scripts/PruebaRotacion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8dcf39a351192fd49a45b57c7294b8ff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
